+++ crates/blockifier/src/blockifier/transaction_executor_test.rs
@@ -1,8 +1,8 @@
 use assert_matches::assert_matches;
 use pretty_assertions::assert_eq;
 use rstest::rstest;
-use starknet_api::felt;
 use starknet_api::transaction::{Fee, TransactionVersion};
+use starknet_api::{declare_tx_args, deploy_account_tx_args, felt, invoke_tx_args};
 use starknet_types_core::felt::Felt;
 
 use crate::blockifier::config::TransactionExecutorConfig;
@@ -13,6 +13,7 @@ use crate::blockifier::transaction_executor::{
 };
 use crate::bouncer::{Bouncer, BouncerWeights};
 use crate::context::BlockContext;
+use crate::nonce;
 use crate::state::cached_state::CachedState;
 use crate::state::state_api::StateReader;
 use crate::test_utils::contracts::FeatureContract;
@@ -39,7 +40,6 @@ use crate::transaction::test_utils::{
 };
 use crate::transaction::transaction_execution::Transaction;
 use crate::transaction::transactions::L1HandlerTransaction;
-use crate::{declare_tx_args, deploy_account_tx_args, invoke_tx_args, nonce};
 
 fn tx_executor_test_body<S: StateReader>(
     state: CachedState<S>,
