+++ crates/blockifier/src/concurrency/fee_utils_test.rs
@@ -1,14 +1,13 @@
 use num_bigint::BigUint;
 use rstest::rstest;
-use starknet_api::felt;
 use starknet_api::transaction::{Fee, ValidResourceBounds};
+use starknet_api::{felt, invoke_tx_args};
 use starknet_types_core::felt::Felt;
 
 use crate::concurrency::fee_utils::{add_fee_to_sequencer_balance, fill_sequencer_balance_reads};
 use crate::concurrency::test_utils::create_fee_transfer_call_info;
 use crate::context::BlockContext;
 use crate::fee::fee_utils::get_sequencer_balance_keys;
-use crate::invoke_tx_args;
 use crate::state::state_api::StateReader;
 use crate::test_utils::contracts::FeatureContract;
 use crate::test_utils::initial_test_state::{fund_account, test_state, test_state_inner};
