+++ crates/blockifier/src/concurrency/versioned_state_test.rs
@@ -12,7 +12,14 @@ use starknet_api::core::{
     PatriciaKey,
 };
 use starknet_api::transaction::{Calldata, ContractAddressSalt, ValidResourceBounds};
-use starknet_api::{calldata, class_hash, contract_address, felt, patricia_key};
+use starknet_api::{
+    calldata,
+    class_hash,
+    contract_address,
+    deploy_account_tx_args,
+    felt,
+    patricia_key,
+};
 
 use crate::abi::abi_utils::{get_fee_token_var_address, get_storage_var_address};
 use crate::concurrency::test_utils::{
@@ -44,7 +51,7 @@ use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::objects::{HasRelatedFeeType, TransactionInfoCreator};
 use crate::transaction::test_utils::{l1_resource_bounds, max_resource_bounds};
 use crate::transaction::transactions::ExecutableTransaction;
-use crate::{compiled_class_hash, deploy_account_tx_args, nonce, storage_key};
+use crate::{compiled_class_hash, nonce, storage_key};
 
 #[fixture]
 pub fn safe_versioned_state(
