+++ crates/blockifier/src/fee/gas_usage_test.rs
@@ -2,6 +2,7 @@ use std::num::NonZeroU128;
 
 use pretty_assertions::assert_eq;
 use rstest::{fixture, rstest};
+use starknet_api::invoke_tx_args;
 use starknet_api::transaction::{EventContent, EventData, EventKey, Fee};
 use starknet_types_core::felt::Felt;
 
@@ -15,7 +16,6 @@ use crate::fee::gas_usage::{
     get_da_gas_cost,
     get_message_segment_length,
 };
-use crate::invoke_tx_args;
 use crate::state::cached_state::StateChangesCount;
 use crate::test_utils::{DEFAULT_ETH_L1_DATA_GAS_PRICE, DEFAULT_ETH_L1_GAS_PRICE};
 use crate::transaction::objects::{FeeType, GasVector, StarknetResources};
