+++ crates/blockifier/src/transaction/transactions_test.rs
@@ -22,7 +22,16 @@ use starknet_api::transaction::{
     TransactionVersion,
     ValidResourceBounds,
 };
-use starknet_api::{calldata, class_hash, contract_address, felt, patricia_key};
+use starknet_api::{
+    calldata,
+    class_hash,
+    contract_address,
+    declare_tx_args,
+    deploy_account_tx_args,
+    felt,
+    invoke_tx_args,
+    patricia_key,
+};
 use starknet_types_core::felt::Felt;
 use strum::IntoEnumIterator;
 
@@ -66,7 +75,6 @@ use crate::test_utils::prices::Prices;
 use crate::test_utils::{
     create_calldata,
     create_trivial_calldata,
-    default_testing_resource_bounds,
     get_syscall_resources,
     get_tx_resources,
     test_erc20_sequencer_balance_key,
@@ -124,9 +132,6 @@ use crate::versioned_constants::VersionedConstants;
 use crate::{
     check_transaction_execution_error_for_custom_hint,
     check_transaction_execution_error_for_invalid_scenario,
-    declare_tx_args,
-    deploy_account_tx_args,
-    invoke_tx_args,
     nonce,
     retdata,
 };
@@ -1557,7 +1562,7 @@ fn test_validate_accounts_tx(
                                                        * the address of
                                                        * faulty_account. */
         contract_address_salt: salt_manager.next_salt(),
-        resource_bounds: default_testing_resource_bounds(),
+        resource_bounds: ValidResourceBounds::create_for_testing(),
         ..default_args
     });
     let error = account_tx.execute(state, block_context, true, true).unwrap_err();
@@ -1573,7 +1578,7 @@ fn test_validate_accounts_tx(
             scenario: GET_BLOCK_HASH,
             contract_address_salt: salt_manager.next_salt(),
             additional_data: None,
-            resource_bounds: default_testing_resource_bounds(),
+            resource_bounds: ValidResourceBounds::create_for_testing(),
             ..default_args
         });
         let error = account_tx.execute(state, block_context, true, true).unwrap_err();
@@ -1588,7 +1593,7 @@ fn test_validate_accounts_tx(
         let account_tx = create_account_tx_for_validate_test_nonce_0(FaultyAccountTxCreatorArgs {
             scenario: GET_SEQUENCER_ADDRESS,
             contract_address_salt: salt_manager.next_salt(),
-            resource_bounds: default_testing_resource_bounds(),
+            resource_bounds: ValidResourceBounds::create_for_testing(),
             ..default_args
         });
         let error = account_tx.execute(state, block_context, true, true).unwrap_err();
@@ -1611,7 +1616,7 @@ fn test_validate_accounts_tx(
             contract_address_salt: salt_manager.next_salt(),
             additional_data: None,
             declared_contract: Some(FeatureContract::TestContract(declared_contract_cairo_version)),
-            resource_bounds: default_testing_resource_bounds(),
+            resource_bounds: ValidResourceBounds::create_for_testing(),
             ..default_args
         },
     );
@@ -1628,7 +1633,7 @@ fn test_validate_accounts_tx(
                 declared_contract: Some(FeatureContract::AccountWithLongValidate(
                     declared_contract_cairo_version,
                 )),
-                resource_bounds: default_testing_resource_bounds(),
+                resource_bounds: ValidResourceBounds::create_for_testing(),
                 ..default_args
             },
         );
@@ -1648,7 +1653,7 @@ fn test_validate_accounts_tx(
                 declared_contract: Some(FeatureContract::AccountWithoutValidations(
                     declared_contract_cairo_version,
                 )),
-                resource_bounds: default_testing_resource_bounds(),
+                resource_bounds: ValidResourceBounds::create_for_testing(),
                 ..default_args
             },
         );
@@ -1664,7 +1669,7 @@ fn test_validate_accounts_tx(
                 contract_address_salt: salt_manager.next_salt(),
                 additional_data: Some(vec![Felt::from(CURRENT_BLOCK_TIMESTAMP_FOR_VALIDATE)]),
                 declared_contract: Some(FeatureContract::Empty(declared_contract_cairo_version)),
-                resource_bounds: default_testing_resource_bounds(),
+                resource_bounds: ValidResourceBounds::create_for_testing(),
                 ..default_args
             },
         );
@@ -1686,7 +1691,7 @@ fn test_validate_accounts_tx(
                     Felt::from(0_u64), // Sequencer address for validate.
                 ]),
                 declared_contract: Some(FeatureContract::Empty(declared_contract_cairo_version)),
-                resource_bounds: default_testing_resource_bounds(),
+                resource_bounds: ValidResourceBounds::create_for_testing(),
                 ..default_args
             },
         );
