+++ crates/starknet_api/src/test_utils.rs
@@ -1,0 +1,3 @@
+pub mod declare;
+pub mod deploy_account;
+pub mod invoke;
