+++ crates/starknet_api/src/transaction.rs
@@ -955,6 +955,12 @@ impl ValidResourceBounds {
             Self::AllResources(AllResourceBounds { l2_gas, .. }) => *l2_gas,
         }
     }
+
+    // TODO(Nimrod): Default testing bounds should probably be AllResourceBounds variant.
+    #[cfg(any(feature = "testing", test))]
+    pub fn create_for_testing() -> Self {
+        Self::L1Gas(ResourceBounds { max_amount: 0, max_price_per_unit: 1 })
+    }
 }
 
 #[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Hash, Ord, PartialOrd, Serialize)]
